# 🔥 Performance: Add keyword search optimization with 15.2x improvement - Issue #7